requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
scikit-learn==1.5.2
joblib==1.4.2
numpy==2.1.3
sqlmodel==0.0.24
python-dotenv==1.0.1
groq==0.4.0
